requirements.txt
openai>=1.40.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
pillow>=10.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
